--- lib/screens/dashboard.dart
@@ -1,115 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:natbank/screens/contact_list.dart';
-import 'package:natbank/screens/transactions_list.dart';
-
-class Dashboard extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: Text('Dashboard')),
-      backgroundColor: Colors.blue[300],
-      body: LayoutBuilder(
-        builder: (context, constrainedBox) => SingleChildScrollView(
-          child: ConstrainedBox(
-            constraints: BoxConstraints(
-              minHeight: constrainedBox.minHeight,
-            ),
-            child: Column(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Image.asset(
-                      'images/natbank_lightBlue500.png',
-                      scale: 0.5,
-                    ),
-                  ),
-                  Container(
-                    height: 120,
-                    child: ListView(
-                      scrollDirection: Axis.horizontal,
-                      children: <Widget>[
-                        FeatureItem(
-                          'Transferir',
-                          Icons.monetization_on,
-                          onClick: () => _showContactsList(context),
-                        ),
-                        FeatureItem(
-                          'Feed de transferências',
-                          Icons.description,
-                          onClick: () => _showTransactionsList(context),
-                        ),
-                        FeatureItem(
-                          'Pagar',
-                          Icons.payment,
-                          onClick: () => print('Pagar foi clicado'),
-                        ),
-                      ],
-                    ),
-                  ),
-                ]),
-          ),
-        ),
-      ),
-    );
-  }
-
-  void _showContactsList(BuildContext context) {
-    Navigator.of(context)
-        .push(MaterialPageRoute(builder: (context) => ContactList()));
-  }
-
-  _showTransactionsList(BuildContext context) {
-    Navigator.of(context)
-        .push(MaterialPageRoute(builder: (context) => TransactionsList()));
-  }
-}
-
-class FeatureItem extends StatelessWidget {
-  final String name;
-  final IconData icon;
-  final Function onClick;
-
-  FeatureItem(
-    this.name,
-    this.icon, {
-    @required this.onClick,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Material(
-        color: Theme.of(context).primaryColor,
-        child: InkWell(
-          // InkWell é o GestureDetector do MaterialDesign para animações
-          onTap: () => onClick(),
-          child: Container(
-            padding: EdgeInsets.all(8.0),
-            width: 150,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: <Widget>[
-                Icon(
-                  icon,
-                  color: Colors.white,
-                  size: 32.0,
-                ),
-                Text(
-                  name,
-                  style: TextStyle(
-                    color: Colors.white,
-                    fontSize: 16.0,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
