+++ lib/screens/dashboard/dashboard.dart
@@ -1,0 +1,98 @@
+import 'package:flutter/material.dart';
+import 'package:natbank/screens/contact_list.dart';
+import 'package:natbank/screens/transactions_list.dart';
+
+import 'app_bar.dart';
+
+class Dashboard extends StatefulWidget {
+  @override
+  _DashboardState createState() => _DashboardState();
+}
+
+class _DashboardState extends State<Dashboard> {
+  bool _showMenu;
+
+  @override
+  void initState() {
+    super.initState();
+    _showMenu = false;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+//    final dependencies = AppDependencies.of(context).user;
+    return Scaffold(
+      backgroundColor: Theme.of(context).primaryColor /*Colors.blue[300]*/,
+      body: SingleChildScrollView(
+        child: Stack(
+          alignment: Alignment.topCenter,
+          children: <Widget>[
+            DashboardAppBar(showMenu: _showMenu, onTap: () {
+              setState(() {
+                _showMenu = !_showMenu;
+              });
+            },),
+          ],
+        ),
+      ),
+    );
+  }
+
+  void _showContactsList(BuildContext context) {
+    Navigator.of(context)
+        .push(MaterialPageRoute(builder: (context) => ContactList()));
+  }
+
+  _showTransactionsList(BuildContext context) {
+    Navigator.of(context)
+        .push(MaterialPageRoute(builder: (context) => TransactionsList()));
+  }
+}
+
+class FeatureItem extends StatelessWidget {
+  final String name;
+  final IconData icon;
+  final Function onClick;
+
+  FeatureItem(
+    this.name,
+    this.icon, {
+    @required this.onClick,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Material(
+        color: Colors.lightBlue[500],
+        child: InkWell(
+          // InkWell é o GestureDetector do MaterialDesign para animações
+          onTap: () => onClick(),
+          child: Container(
+            padding: EdgeInsets.all(8.0),
+            width: 130,
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                Icon(
+                  icon,
+                  color: Colors.white,
+                  size: 32.0,
+                ),
+                Text(
+                  name,
+                  style: TextStyle(
+                    color: Colors.white,
+                    fontSize: 16.0,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
