+++ pubspec.yaml
@@ -58,6 +58,7 @@ flutter:
     - images/natbank_lightBlue.png
     - images/natbank_lightBlue400.png
     - images/natbank_lightBlue500.png
+    - images/logo_letters.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
