+++ test/matchers/matchers.dart
@@ -1,6 +1,6 @@
 // can be used to find others FeatureItems such as Pagar e Feed de Transferências
 import 'package:flutter/material.dart';
-import 'package:natbank/screens/dashboard.dart';
+import 'package:natbank/screens/dashboard/dashboard.dart';
 
 bool featureItemMatcher(Widget widget, String name, IconData icon) {
   return widget is FeatureItem && widget.icon == icon && widget.name == name;
